requirements.txt
alembic==0.9.9
click==6.7
Flask==1.0.1
Flask-SQLAlchemy==2.3.1
flask-login
flask-user
Flask-Session
WTForms-Alchemy
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pytest==3.5.1
pytest-flask==0.10.0
Werkzeug==0.14.1
psycopg2

